requirements.txt
flask
flask_login
flask_sqlalchemy
flask_wtf
flask_bcrypt
pyotp
pycryptodome